Add methods necessary for condvar implementation
Adds `MutexGuard::mutex` which returns the associated mutex.

Also adds `RwLock::reader_count_acquire` and
`RwLock::writer_count_acquire` so that `RwLock::try_read` and
`RwLock::try_write` (in Theseus) don't spuriously fail. This isn't
strictly necessary, but it was necessary for `Mutex::try_lock` (in
Theseus) to not spuriously fail and it would make sense for that
behaviour to be consistent across sync primitives.

Also this commit only modifies the spin mutex so building with the fair
or ticket features enabled will result in a compile error. Since Theseus
only uses the spin implementation, and these changes won't be upstreamed
I figured there's no point of changing the other mutexes.

Signed-off-by: Klimenty Tsoutsman <klim@tsoutsman.com>